my-project/blog/requirements.txt
flask==2.1.3
flask-sqlalchemy==2.5.1
flask-security-too==4.1.4
flask-restful==0.3.9
bcrypt
email_validator
flask_mail
redis
celery[redis]